Works on all devices without crashing
I optimized all images so that they are at the minimal file size and it
works on my ipod without crashing. Also, I deleted a few lines of code
that may have possibly caused the game to crash. Maybe if I implement
those now, the memory will be enough to sustain it, though.