Hid the HUD and initialized toolbar after play is clicked
To Do: I have to Change the title graphic, I have to edit the coins and
make them burps, I also need to add a simple tutorial with a BOOL. I
need to edit the Missions node and make the knight move a little
faster. I have to change the graphics for the pause, and quit buttons,
and add a restart button to the pause layer. Find an alternative to the
CCParticleEffects. Change Store graphics and change the coin system to
burps (or farts) and have the people that want to buy things buy with
burps.