Fixed toolbar and scaled knight
So now, I just have to scale and optimizwe everything for the iPad and
iPad retina and make the Game Play work. For example, the knight should
be able to collect the foods, so I need to alter the algorithm for
generation. Also, tomorrow, I need to add a ton of missions and make
them all autogenerated. Make some ridiculous missions like in Jetpack
Joyride :)